[DMS-183] Move tests into general folder
